refactor: event -> promotion
