+++ src/main/kotlin/christmas/domain/promotion/Promotion.kt
@@ -1,10 +1,10 @@
-package christmas.domain.event
+package christmas.domain.promotion
 
 import christmas.domain.money.Money
 import christmas.domain.order.Orders
 import christmas.domain.product.Product
 
-interface Event {
+interface Promotion {
     fun discountAmount(orders: Orders): Money
     fun giveawayProduct(orders: Orders): Product?
 }
