+++ src/main/kotlin/christmas/domain/promotion/WeekendMainDishDiscount.kt
@@ -1,13 +1,13 @@
-package christmas.domain.event
+package christmas.domain.promotion
 
-import christmas.domain.event.util.DayChecker.isWeekday
 import christmas.domain.money.Currency
 import christmas.domain.money.Money
 import christmas.domain.order.Orders
 import christmas.domain.product.Product
 import christmas.domain.product.ProductType
+import christmas.domain.promotion.util.DayChecker.isWeekend
 
-class WeekdayDessertDiscount : Event {
+class WeekendMainDishDiscount : Promotion {
     companion object {
         private val DISCOUNT_AMOUNT = Money.longValueOf(2023, Currency.KRW)
     }
@@ -15,9 +15,9 @@ class WeekdayDessertDiscount : Event {
     override fun discountAmount(orders: Orders): Money {
         val dayOfWeek = orders.date.dayOfWeek
 
-        if (dayOfWeek.isWeekday()) {
-            val dessertCount = countDesserts(orders)
-            val totalDiscountMoney = DISCOUNT_AMOUNT.multiply(dessertCount)
+        if (dayOfWeek.isWeekend()) {
+            val mainCount = countMainDishes(orders)
+            val totalDiscountMoney = DISCOUNT_AMOUNT.multiply(mainCount)
             return totalDiscountMoney
         }
 
@@ -28,9 +28,9 @@ class WeekdayDessertDiscount : Event {
         return null
     }
 
-    private fun countDesserts(orders: Orders): Int {
+    private fun countMainDishes(orders: Orders): Int {
         return orders.orders
-            .filter { it.product.productType == ProductType.DESSERT }
+            .filter { it.product.productType == ProductType.MAIN }
             .sumOf { order -> order.amount }
     }
 }
